Spawning trees and rocks
Trees and rocks now randomly generate on grass terrain. there is also a chance to get grass terrain to spawn with only rocks or trees on the ends of it. The log in the water was moved slightly higher above water to help with the chicken dyeing from water(still needs to be tested out).